stt_server/requirements-lock.txt
--extra-index-url https://download.pytorch.org/whl/cu118
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.4.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asteroid-filterbanks==0.4.0
async-timeout==4.0.3
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cmake==3.31.10
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.0
cryptography==41.0.7
cycler==0.12.1
docopt==0.6.2
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.121.3
filelock==3.19.1
fonttools==4.60.2
frozenlist==1.8.0
fsspec==2025.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.1.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httplib2==0.31.2
huggingface-hub==0.20.3
HyperPyYAML==1.2.3
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
julius==0.2.7
kenlm @ https://github.com/kpu/kenlm/archive/master.zip#sha256=f935c2ac0be6dd1bf96a5852973dad6103c48e866cf33532459fb1336bea4bfe
kiwisolver==1.4.7
lightning==2.6.0
lightning-utilities==0.15.2
Mako==1.3.11
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.2.1
ninja==1.13.0
noisereduce==3.0.3
numpy==1.25.2
omegaconf==2.3.0
optuna==4.8.0
packaging==26.1
pandas==2.3.3
pillow==11.3.0
primePy==1.3
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==15.0.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scikit-learn==1.6.1
scipy==1.13.1
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.49
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.5
threadpoolctl==3.6.0
tomli==2.4.1
torch==2.7.1+cu118
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1+cu118
torchmetrics==1.8.2
tqdm==4.67.3
typer==0.23.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.38.0
yarl==1.22.0
zipp==3.23.1
